refactor: 실제 사용되는 DB와 관련된 모듈은 infra 패키지로 이동
